Move in_memory storage to its own module
